UpdateAPI function
This function checks for user info stored in local storage, if found it create the APIHandler class, if no user data is stored it redirects the user to the login page. I also added a query string for previous page maybe I will use this to bring the user back to the last page they were on.